backend/requirements.txt
Django==6.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
gunicorn==25.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
psutil==6.1.1
django-cors-headers==4.9.0
python-json-logger==3.2.1
whitenoise==6.9.0
